Revert to using the SDL2 WindowSurface shortstop
I found out after doing some of the texture work that SDL2 still has the
idea of a screen surface if you choose to forgo practically all texture
usage.

Ultimately, using textures is preferable. This branch is here as an
alternative if it turns out to not be feasible to do

I think everything that matters works here (read, not CD audio)